Add check for duplicated IDs in the input FASTA
Previously, there was a check that looked for overlap between all ASVs in the input FASTA and the feature table, but there were downstream errors that could be caused by having duplicated sequence IDs in the input fasta. This fix adds a check that no sequence IDs in the input FASTA appear more than once.